Update destructor logic for HashMaps [#141511349]
In previous implementation of hash maps, all the bucket chains have been
traversed no matter if the bucket is pointing to a chain or not.

In this new version of `Clear()` we keep track of the indices for every
bucket where a chain is created and we simply deallocated these chains.
Therefore, no time is wasted on checking unallocated bucket chains.

Signed-off-by: Omer Arap <oarap@pivotal.io>