Remove dead code
We never send null part constraints from the relacache translator
hence we do not handling for the same.

This code was probably added to support the older minidumps. There are
a few very old minidump files which do not contain the part constraint
tag in relation tag.

Now with the fix on the relcache translator side in GPDB, the only case
when we send NULL part constraints is when there are no indices and no
default partitions; we still don't need null check for part constraint in
this case because the `fDummyConstraint` will always be true.

[Ref #149769559]

Signed-off-by: Dhanashree Kashid <dkashid@pivotal.io>