Fix Orca crash if a scalar subquery appears in the outer child of IN subquery
Prior to this commit, the predicate of the subquery was always assumed
to be scalar and only the scalar child in the expression was being
extracted. But, in case the predicate is a subquery, the extracted
scalar child references to a column which was previosly part of the
subquery but is dropped as only the scalar is getting added. This commit
takes care of checking if the predicate is a subquery and returns the
entire plan which ensures the plan generation in ORCA.

Signed-off-by: Dhanashree Kashid <dkashid@pivotal.io>