Fix Orca crash if a scalar subquery appears in the outer child of IN subquery
Prior to this commit, ORCA constructed an incorrect join predicate while
converting the IN `CScalarSubqueryAny` to Semi Join. During this
process, ORCA extracts the predicate for the outer child and asumes to be
always a scalar child; for cases where it can be a subquery, ORCA just dropped
the subquery/subtree which produced the scalar column in the predicate.
Hence failing the assert in `CScalarSubquery.cpp`.
This commit takes care of checking if the predicate is a subquery and
returns the entire subquery tree.

Signed-off-by: Dhanashree Kashid <dkashid@pivotal.io>