Skip generation of multi-stage agg alternatives for non-scalar DQA on distribution column
Previously, for queries containing a non-scalar DQA(Distinct Qualified
Aggregate) on the distribution column as well as non-splittable
aggregate functions, we generated multi-stage aggregate alternatives
which was unnecessary, as these would never be chosen--either due to
costing or correctness. A non-scalar DQA is one which is calculated for
more than one group, usually through the specification of a GROUP BY
clause.

With this commit, we avoid generating such alternatives for queries of
this kind.

Signed-off-by: Melanie Plageman <mplageman@pivotal.io>