Remove Hungarian notation from libnaucrates and libgpos
Hungarian notation removal: GPOS and Naucrates libraries

This is the first commit of the initiative to remove Hungarian notation
and replace it with a new convention -- a custom convention based on
common C++ naming conventions. The style guide for this convention is
forthcoming. Look for an updated README with details in future commits.

 Co-authored-by: Melanie Plageman <mplageman@pivotal.io>
 Co-authored-by: Ekta Khanna <ekhanna@pivotal.io>
 Co-authored-by: Bhuvnesh Chaudhary <bchaudhary@pivotal.io>
 Co-authored-by: Abhijit Subramanya <asubramanya@pivotal.io>
 Co-authored-by: Sambitesh Dash <sdash@pivotal.io>
 Co-authored-by: Omer Arap <oarap@pivotal.io>