Fixing crash while deriving relation properties for index apply alternatives
For Xform `CXformInnerJoin2DynamicBitmapIndexGetApply`, the pattern of the input expression
was expected to be:
```
+--CLogicalInnerJoin
|--CPatternLeaf
|--CLogicalDynamicGet
+--CPatternTree
```

Consider an input expression of the form
```
+--CLogicalInnerJoin
|--CLogicalInnerJoin
|  |--CLogicalGet
|  |--CLogicalGet
|  +--CScalarCmp
|--CLogicalDynamicGet "foo_prt"
+--CScalarCmp (=)   origin: [Grp:4, GrpExpr:0]
```

In PexprExtract, given the input pattern, it extracts the below expression
```
+--CLogicalInnerJoin
|--CLogicalInnerJoin    (1) --> We did not extract children of this join
|--CLogicalDynamicGet
+--CScalarCmp (=)
   |--CScalarIdent
   +--CScalarIdent
```

Notice that for CLogicalInnerJoin (1), the above tree does not contain
information of the child. However, CXformInnerJoin2DynamicBitmapIndexGetApply
creates a CTE Producer expression (for partial indexes) and performs expression
preprocessing.  During expression preprocessing, while deriving the relational
properties of the CLogicalInnerJoin (1), it expects the entire tree. But since
its not extracted, it fails assertion in debug build or crashes in retail build

Since, CTE expression must be preprocessed, the entire tree must exist, and
instead of CPatternLeaf, the expected pattern should be CPatternTree. This
commit fixes the issue by updating the expected pattern to CPatternTree.

However, for deep trees, PexprExtract may extract lot of expressions and could
increase the optimization time of the queries.  So, the xform should only be
applied once. This commit marks the
`CXformInnerJoin2DynamicBitmapIndexGetApply` as apply once and all the derived
xforms of the base class `CXformJoin2IndexApply` as apply once.
Additionally, it also refactors the code to handle such apply once expression
by making it a xform property instead of util function.

Bump ORCA Version to 2.65