Improve bitmap index matching to better leverage the multi-key indexes
Do not break down a predicate that contains conjnucts with children of the form:
<Ident> op <const> or <Ident> op <const-array> or <boolean column> or
NOT <boolean column>.
Instead treat the predicate as a whole while finding a suitable index
match.  Also, for such predicates, perform the lookup again for the
residual to minimize the number of final residuals and to leverage available
indexes on them.
If there are multiple index options available for the given predicate
then an Index covering maximum predicate columns is selected.

Example, for schema t(a, b, c, f, e)
indexes: i1 on (a), i2 on (e), i3 on (a, b) and i4 on (b, c, a)
and Predicate: (a = 1) AND (b = 2) AND (e = 3) AND (c = 4)

Previously, we separated the children of AND and performed the lookup
independently for each of them; also the lookup stopped at first
matching index hence if there were multiple index matches available for
a child, it would pick the one in the oid order. For instance, for (a =
1) if i1 was first in the list then it would get picked without going
through rest of the indexes which could be potential matches too. The
final index plan would look like:

BitmapTableScan
   Recheck: (a = 1) AND (b = 2) AND (e = 3)
   Filter: c = 4
      |_
         BitmapAnd
	    |_
              BitmapIndexScan on I1
	         Index Cond: a = 1
	    |_
	      BitmapIndexScan on I3
	         Index Cond: b = 2
            |_
	      BitmapIndexScan on I2
	         Index Cond: e = 3

After this commit, the lookup will go through all the available indexes
to pick an optimal index. It will produce:

BitmapTableScan
   Recheck: (a = 1) AND (b = 2) AND (e = 3)
   Filter: c = 4
      |_
         BitmapAnd
	    |_
	      BitmapIndexScan on I3
	         Index Cond: (a = 1) AND (b = 2)
            |_
	      BitmapIndexScan on I2
	         Index Cond: e = 3

If predicate is a disjunct tree, then algorithm behaves same as before
wherein the disjunct tree is broken down and index lookup is performed
on individual children. Additionally in case of join2Index transforms,
the matching algorithm remains same.

In nutshell, the new and improved index matching algorithm is only
specialized for predicates of the form with a schema containing
multi-key indexes (which are very common) :
SELECT * FROM t1 WHERE c = 10 AND b = 40 AND g = 35 AND e = 20 AND h = 50;

Signed-off-by: Dhanashree Kashid <dkashid@pivotal.io>